requirements.txt
pytube
ytmusicapi
pydub
music_tag
wget
pillow
